+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## [Unreleased] 2021-11-22
+## [Unreleased]
+
 - Upgrade to [npm-groovy-lint](https://www.npmjs.com/package/npm-groovy-lint) v9.0.0
   - Upgrade to [CodeNarc 2.2](https://github.com/CodeNarc/CodeNarc/blob/master/CHANGELOG.md)
   - Adds [Groovy 3.0.9](http://groovy-lang.org/changelogs/changelog-3.0.9.html) support.